chore: rename file to follow convention
