+++ app/(fallback)/_layout.tsx
@@ -7,7 +7,7 @@ import React, { useCallback, useState } from 'react';
 import { BottomNavigation } from '@/base/components/bottom-navigation';
 import { YStack } from 'tamagui';
 import GlobalLoadingFallback from '@/base/providers/global-loading-fallback';
-import AuthGuardProvider from '@/auth/providers/AuthGuardProvider';
+import AuthGuardProvider from '@/auth/providers/auth-guard-provider';
 import { Toast } from '@tamagui/toast';
 import { ToastProvider, ToastViewport } from '@tamagui/toast';
